requirements.txt
Flask==2.3.2               # Фреймворк для работы с вебхуками
python-dotenv==1.0.0       # Для работы с переменными окружения из .env
psycopg2~=2.9.10
pyTelegramBotAPI==4.12.0   # Библиотека для работы с Telegram Bot API
gunicorn==21.2.0           # WSGI сервер для запуска Flask-приложений


PyJWT~=2.9.0
requests~=2.32.3
psycopg2-binary~=2.9.10
aiofiles~=24.1.0
httpx~=0.27.2